+++ prt/tables.hpp
@@ -21,3 +21,56 @@ static const U8 WRT_wrd1[256]={
  192,192,192,192,192,192,192,192, 192,192,192,192,192,192,192,192,
  192,192,192,192,192,192,192,192, 192,192,192,192,192,192,192,192,
  };
+ 
+static const U8 AsciiGroupC0[254] ={
+  0, 10,
+  0, 1, 10, 10,
+  0, 4, 2, 3, 10, 10, 10, 10,
+  0, 0, 5, 4, 2, 2, 3, 3, 10, 10, 10, 10, 10, 10, 10, 10,
+  0, 0, 0, 0, 5, 5, 9, 4, 2, 2, 2, 2, 3, 3, 3, 3, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10,
+  0, 0, 0, 0, 0, 0, 0, 0, 5, 8, 8, 5, 9, 9, 6, 5, 2, 2, 2, 2, 2, 2, 2, 8, 3, 3, 3, 3, 3, 3, 3, 8, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10,
+  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 8, 8, 8, 8, 8, 5, 5, 9, 9, 9, 9, 9, 7, 8, 5, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 8, 8, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 8, 8, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10
+};
+static const U8 AsciiGroup[128] = {
+  0,  5,  5,  5,  5,  5,  5,  5,
+  5,  5,  4,  5,  5,  4,  5,  5,
+  5,  5,  5,  5,  5,  5,  5,  5,
+  5,  5,  5,  5,  5,  5,  5,  5,
+  6,  7,  8, 17, 17,  9, 17, 10,
+  11, 12, 17, 17, 13, 14, 15, 16,
+  1,  1,  1,  1,  1,  1,  1,  1,
+  1,  1, 18, 19, 20, 23, 21, 22,
+  23,  2,  2,  2,  2,  2,  2,  2,
+  2,  2,  2,  2,  2,  2,  2,  2,
+  2,  2,  2,  2,  2,  2,  2,  2,
+  2,  2,  2, 24, 27, 25, 27, 26,
+  27,  3,  3,  3,  3,  3,  3,  3,
+  3,  3,  3,  3,  3,  3,  3,  3,
+  3,  3,  3,  3,  3,  3,  3,  3,
+  3,  3,  3, 28, 30, 29, 30, 30
+};
+
+static const U32 WRT_mpw[16]= { 4, 4, 3, 2, 2, 2, 1, 1,  1, 1, 1, 1, 0, 0, 0, 0 };
+static const U8 WRT_mtt[256]= {
+ 0,4,4,5,0,5,3,7, 0,2,4,0,0,0,0,0,
+ 0,0,0,0,0,0,0,0, 1,0,0,0,0,0,0,0,
+ 2,4,3,3,4,6,6,7, 5,4,0,4,3,4,0,1,
+ 1,1,1,1,1,1,1,1, 1,1,7,3,3,4,3,4,
+ 
+ 0,5,5,5,5,5,5,5, 5,5,5,5,5,5,5,5,
+ 5,5,5,5,5,5,5,5, 5,5,5,6,4,3,4,4,
+ 0,5,5,5,5,5,5,5, 5,5,5,5,5,5,5,5,
+ 5,5,5,5,5,5,5,5, 5,5,5,4,4,4,4,4,
+
+ 6,6,6,6,6,6,6,6, 6,6,6,6,6,6,6,6,
+ 6,6,6,6,6,6,6,6, 6,6,6,6,6,6,6,6,
+ 6,6,6,6,6,6,6,6, 6,6,6,6,6,6,6,6,
+ 6,6,6,6,6,6,6,6, 6,6,6,6,6,6,6,6,
+ 6,6,6,6,6,6,6,6, 6,6,6,6,6,6,6,6,
+
+ 7,7,7,7,7,7,7,7, 7,7,7,7,7,7,7,7,
+ 7,7,7,7,7,7,7,7, 7,7,7,7,7,7,7,7,
+ 7,7,7,7,7,7,7,7, 7,7,7,7,7,7,7,7,
+};
+
+static const U32 tri[4]={0,4,3,7}, trj[4]={0,6,6,12};
